requirements.txt
numpy>=1.20.0
torch>=1.8.0
librosa>=0.8.1
ffmpeg-python>=0.2.0
python-dotenv>=0.19.0
scipy>=1.7.0  # Required by librosa
soundfile>=0.10.3  # Required for audio I/O
moviepy>=1.0.3  # For robust video processing
silero-vad>=5.0.0  # For voice activity detection